Fix weird edge case in PhantomJS
When the function.bind polyfill is used on console methods (e.g console.log),
it throws "instanceof called on an object with an invalid prototype property".

This fix will ignore such an error, and assume that the call is not of the form
`new bound()`, which is a weird anyway.