major child_process refactor:
`run` now accepts a block, and will run it while also waiting for the process to end.
retraction will still cause the process to be killed, as will an error
thrown from `block`.

In addition, `stdio` options have been extended:

 - there are now explicit `string` and `buffer` options
 - `pipe` actually gives you a pipe now, as long as you pass a `block`
 - a Stream can be passed to `stdin`
 - new shorthands have been introduced, and default nodejs shorthands
   modified to inherit `stderr` unless explicitly silenced.

Errors thrown for a failed command now include `stderr` output (if it's a
string), as well as the command & arguments that failed.